+++ examples/demo-rollup/tests/bank/zk_rollup/bank_periodic_da_tests.rs
@@ -35,7 +35,7 @@ async fn flaky_bank_tx_tests_periodic_da() -> anyhow::Result<()> {
     )
     .with_zkvm_host_args(mock_da_risc0_host_args())
     .set_config(|c| {
-        c.rollup_prover_config = RollupProverConfig::Skip;
+        c.rollup_prover_config = Some(RollupProverConfig::Skip);
     })
     .start()
     .await?;