+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2025-01-28
 - #2304 removes some mandatory arguments to `RollupBuilder::new` in favor of builder methods.
+- #2310 allows to completely disable prover background test for in `RollupBuilder` for more precise testing.
 ## 2025-01-23
 - #2266 Add soak testing crate for long running tests
 ## 2025-01-22